adding actual exports from index.d.ts
