+++ types/graphql-schema-linter/lib/validation_error.d.ts
