+++ types/graphql-schema-linter/lib/validator.d.ts
@@ -1,0 +1,7 @@
+import { Configuration, Rule } from './configuration';
+import { ValidationError } from './validation_error';
+export function validateSchemaDefinition(
+  schemaDefinition: string,
+  rules: ReadonlyArray<Rule>,
+  configuration: Configuration,
+): ReadonlyArray<ValidationError>;
